requirements.txt
python-telegram-bot==22.5
requests==2.31.0
python-dotenv==1.0.0
aiohttp==3.9.1
pytz==2023.3
anthropic==0.34.2
